LayoutTests/http/tests/security/dataURL/xss-DENIED-from-javascript-url-window-open-expected.txt

CONSOLE MESSAGE: line 1: TypeError: Undefined value
Opener Frame

PASS: Access from a window opened with a data: URL was denied.
